requirements.txt
Flask==2.3.3
Flask-WTF==1.1.1
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Migrate==4.0.4
Flask-Bcrypt==1.0.1
Flask-CORS==3.0.10
SQLAlchemy==1.4.35
mysql-connector-python==8.0.33
email-validator==1.3.1
Werkzeug==2.3.7
python-dotenv==1.0.0
